Added modular health bar to boss (and make boss actually work)
Added export variable for entity to make health bar modular
Removed nonmodular implementation for player health bar

Added functioning health bar to boss as a HUD element to the boss_level scene.
Made boss set it's current hp in the ready function since otherwise it will take the default value of the export value even if it is changed.

Added a RayCast2D to the Boss so it can exit the charge state when it hits a wall. It is under the Sprite2D so it will flip with where the boss is facing.

The charge state now grants super armor to the boss.

Added a short, middle platform to the boss level.